- init socket

+++ src/pages/Home/index.js
@@ -2,7 +2,11 @@ import {useState, useEffect} from "react";
 import {useDispatch, useSelector} from "react-redux";
 import i18n from 'i18next'
 
+import useSocket from 'hooks/useSocket';
+
 import {gameType} from "constant/config";
+
+import checkCmd from "helpers/checkCmd";
 
 import {setSettings} from "store/actions/settingsAction";
 
@@ -12,6 +16,7 @@ import ROULETTE from "games/ROULETTE";
 import KENO from "games/KENO";
 import DOGS_6 from "games/DOGS_6";
 import HORSES_8_VR from "games/HORSES_8_VR"
+
 import Loader from "components/Loader";
 import Nav from "components/Nav";
 import Betslip from "modules/Betslip";
@@ -40,22 +45,39 @@ const setGame = (id) => {
 }
 
 const Home = () => {
+    const { sendMessage } = useSocket()
     const dispatch = useDispatch()
     const [loading, setLoading] = useState(true)
     const {notification} = useSelector((state) => state.notification)
     const {game} = useSelector((state) => state.game)
+    const {socket, receivedMessage} = useSelector((state) => state.socket);
 
     useEffect(() => {
-        dispatch(setSettings()).then((json) => {
-            if (json.hasOwnProperty('data')) {
-                sessionStorage.clear()
-            }
-            else {
-                i18n.changeLanguage(json.account.language || 'en');
-                setLoading(false)
-            }
-        })
-    }, []);
+        // console.log(socket)
+
+        if (socket && socket.readyState === WebSocket.OPEN) {
+            sendMessage({cmd:`account/${sessionStorage.getItem('authToken')}/settings`})
+        }
+        else {
+            dispatch(setSettings()).then((json) => {
+                if (json.hasOwnProperty('data')) {
+                    sessionStorage.clear()
+                }
+                else {
+                    i18n.changeLanguage(json.account.language || 'en');
+                    setLoading(false)
+                }
+            })
+        }
+    }, [socket]);
+
+    useEffect(() => {
+        if (receivedMessage !== '' && checkCmd('settings', receivedMessage.cmd) && receivedMessage.hasOwnProperty('username')) {
+            dispatch(setSettings(receivedMessage))
+            i18n.changeLanguage(receivedMessage.language || 'en');
+            setLoading(false)
+        }
+    }, [receivedMessage])
 
     return (
         <div className={style.block}>
@@ -65,16 +87,23 @@ const Home = () => {
                     <Loader />
                  :
                     <>
-                        {
-                            game &&
-                            <div className={style.decor}>
-                                <img
-                                    src={`/img/decor/${game.type}.jpeg`}
-                                    alt="Decor"
-                                />
-                            </div>
-                        }
+                        {/*{*/}
+                        {/*    game &&*/}
+                        {/*    <div className={style.decor}>*/}
+                        {/*        <img*/}
+                        {/*            src={`/img/decor/${game.type}.jpeg`}*/}
+                        {/*            alt="Decor"*/}
+                        {/*        />*/}
+                        {/*    </div>*/}
+                        {/*}*/}
                         <Nav />
+                        <button
+                            onClick={() => {
+                                socket.close()
+                            }}
+                        >
+                            Disconnect
+                        </button>
                         <div className={style.content}>
                             <div className={style.column}>
                                 <div className={style.banners}>
